prov/sockets: terminate incoming discarded messages properly
The sockets provider may choose to discard an incoming message.  For
example, if an originator initiates an RMA write to a target, but the
target generates a permission denied error, the sockets provider needs
to consume and discard all the as-yet-unread data on the incoming data
socket.

sock_discard_field() is invoked in these cases.  It reads the next
chunk of data from the socket, discards it, and adjusts the count for
how many bytes remain to be discarded.

If all other processing has completed when it consumes all remaining
incoming bytes of the to-be-discarded message, it sets
pe_entry->is_complete to 1, and additional processing outside of
sock_discard_field() moves on to read the next message.

However, if the entire message has been consumed (i.e.,
pe_entry->rem==0), but other processing has *not* completed (i.e.,
pe_entry->done_len != pe_entry->total_len), then we exit
sock_discard_data() without setting pe_entry->is_complete to 1.
Eventually, when the socket is polled again for incoming data,
sock_discard_data() will be called again, but it would exit early when
it saw that there was no more incoming data to be discarded.

Specifically: pe_entry->is_complete was never set to 1, and the
sockets provider never moved on from trying to discard this incoming
message.

This commit changes the "exit early" condition in sock_discard_field()
to also check if the additional dicard processing is done (i.e., the
done_len==total_len check) so that it can set pe_entry->is_complete to
1 if everything is done done done and we have totally finished
discarding.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>