Add capability bits to handle communication scope.
Add two new capability bits, FI_COMM_LOCAL and FI_COMM_REMOTE, to handle
communication scope. A single bit is not enough to express that a
provide support solely local communication (such as shared memory) or
solely remote communication (such as usNIC).

This will affect the API but not the ABI.

Signed-off-by: Ben Turrubiates <bturrubi@cisco.com>