Add communication scope to domain attribute.
Allow the application to provide the type of communication it requires
as a hint. This implementation does not use extra bits that may
otherwise be used by the capabilities or modes. By providing flags
specific to the domain attribute, this allows more flexibility if it is
determined that FI_COMM_LOCAL and FI_COMM_REMOTE do not offer the
granularity required.

This will affect both the API and the ABI.

Signed-off-by: Ben Turrubiates <bturrubi@cisco.com>