util: Add communication map abstraction.
This would enable providers like RXM to keep track of an underlying provider's
connection oriented endpoints info. This might also be used by the upcoming shared
memory provider.

A connection map (cmap) stores handles which have a one-to-one correpondence to the
addresses in an AV. A peer list is used store incoming connections info for remote
endpoints whose fabric addresses aren't added to the AV. A cmap handle also stores
a key which can be sent on message headers to remote endpoints to implement FI_SOURCE.

Signed-off-by: Arun C Ilango <arun.ilango@intel.com>