man: Clarify the completion ordering of multi-recv operation
A single posted recv with FI_MULTI_RECV flag set can generate
multiple completion events. The ordering of these completions was
not clearly defined. Application may need to take extra effort to
determine the ordering of received data, an overhead that can be
avoided if the ordering is enforced.

Add the text to clarify that strict ordering is guaranteed for
completions generated from a single multi-recv operation.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>