usnic: Verify domain name hint.
Previously it seems that the usNIC provider was not checking the domain
name hint. This needs to be validated in both getinfo and in fi_fabric.

Since the domain name was NULL in pre-1.4 releases of libfabric, the
comparison logic is somewhat complicated:

1. If version is greater than or equal to 1.4 and a non-NULL name hint
   is provided, then compare the hint and reference domain name.
2. If the version is greater than or equal to 1.4 and the provider hint
   is NULL, then treat this as valid. This could be an application
   requesting a 1.4 domain and not explicitly providing a name hint.
3. If the version is less than 1.4 and a name hint is provided, then
   this is always going to be invalid as the name should always be NULL
   in pre-1.4.
4. If the version is less than 1.4 and the name hint is NULL, then this
   will always be valid as all domains have NULL as their name in
   pre-1.4.

If there is no version provided, and a name hint is provided then
version 1.4 is assumed. This is the case when used from fi_fabric.

Signed-off-by: Ben Turrubiates <bturrubi@cisco.com>