util/pingpong.c: Fix handling of src/dst port.
The -B and -P options were not being used before. These ports should be
used when establishing the control socket communication paths. The
server will default to a source port of 47592 as before and the client
will default to connecting to 47592. If provided, the server will bind
to the port given with -B. If provided, the client will bind to the port
given with -B locally and connect to the port provided with -P (or the
default).

Signed-off-by: Ben Turrubiates <bturrubi@cisco.com>