fi_fini destructor attribute refactoring
on Unix-like systems there is way to register callback
on module unload/process termination. for windows platforms
there is no way to do this for "C" module. to workaround
this issue such "destructors" are called from DllMain
function when module is unloaded.

so, change is: wrap fi_fini function to FI_DESTRUCTOR macro
which expanded as __attribute__((destructor)) on Unix
systems and "void fi_fini" on windows systems to be called
from DllMain function

Signed-off-by: Oblomov, Sergey <sergey.oblomov@intel.com>