man/cm: Document that EPs cannot be reused after a failed connection
If a connection request fails, the associated EP must be destroyed.
It cannot be re-used.

The reason is that a significant amount of connection state may be
maintained by the provider, including corresponding kernel agents.
Once the connection request fails or is rejected, that state may
not be reset.  Preventing the EP from being re-used in a new
connection request avoids the provider needing to maintain a
sigificant amount of state information, such as posted messages,
that would be needed to rebuild the EP in order for it to be
reused.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>