prov/udp: Reduce default cq, tx, rx counts
Some systems limit each user to only 1024 open files.  If a CQ is
opened with an fd wait object, the default of 64k will be too
large.  Reduce the default size to 1k.  Similarly reduce the rx/tx
context counts.

This will allow the udp provider to pass the cq_size test.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>