prov/psm,psm2: Fix the return value of fi_cq_readfrom
There is a corner case that fi_cq_readfrom needs to return 0 --
when the input count is 0 and the CQ is not empty. This is useful
as a progress call that also checks the CQ state: return value
-FI_EAGAIN or 0 indicates whether the CQ is empty or not. The util
poll function relies on such behavior.

The conditions for identifying the corner case was not exactly
right in the existing code. As the result 0 could be returned
incorrectly under rare cases when multithreading is enabeld. This
patch fixes the conditions.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>