fi_cq_test: Convert error into warning
The CQ test attempts to open cq_cnt CQs, and errors if that fails.
This is resulting in travis errors as a result of open fd limits.
Because the cq_cnt is intended to reflect hardware limits, and not
system software limits, instead of having the test error, simply
report a warning and continue with additional tests.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>