fabric: Add api version field to fi_fabric_attr
There's no way to know what version of the API the application
wants to use when opening a fabric object.

As reported:

"We were not able to find a suitable solution because some
API calls do not have a way to propagate the application-
requested version information to the provider implementation.

Specifically, the fi_getinfo() API function has the uint32_t
version param, which allows propagation of the app-requested
version number down into the provider. But the fi_fabric()
API function does not have any access to versioning information
(i.e., no version parameter, and no access to an info object
returned via fi_getinfo())."

The simplest fix that requires no changes to the app is to
add the version to the fabric attribute.

Fixes #2182

Note: An 'ideal' solution would be to add the version field
directly to fi_info.  However, fi_info isn't available when
calling fi_fabric.  We could change the fi_fabric API to take
fi_info, but that would require applications to modify their
source code.  And if the app wants to run against both newer
and older versions of libfabric, then it would need to carry
both calls and adjust, and, well, it just becomes a mess.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>