usnic: Pad the getinfo extension structures.
The introduction of version 2 of the fi_usnic_info extension in the 1.3
release caused a compatibility issue with older versions of the
extension. This commit adds padding to the structure to explicitly
maintain the alignment requirements for the extension in the 1.3
release. Add explicit padding to all structures used in the getinfo
extension and add the packed attribute.

Relevant documents have been updated to document the issue and recommend
possible solutions.

Signed-off-by: Ben Turrubiates <bturrubi@cisco.com>