fabric/av: Define mechanism to report synchronous errors
See issue #2229.  Introduce a mechanism to report which
addresses failed to be inserted and why when calling
fi_av_insert* as part of a synchronous operation.

Add a new flag value, FI_SYNC_ERR, which indicates that
the caller will pass in an array of integers as the
context parameter.  The provider will write error values
to this array when an error occurs.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>