fabric: Deprecate the sizeleft function
This call is difficult to implement with some providers without
negatively impacting performance or wasting resources.  The
standard way for apps to deal with busy resources is to check
for FI_EAGAIN on data transfer operations.  This is similar to
the socket API.

Remove documentation to these calls and mark the functions
as deprecated.

Fixes #2459

Signed-off-by: Sean Hefty <sean.hefty@intel.com>