fabric: Define behavior if CQ/EQ is overrun by events
We require some sort of detection by the provider to detect
if a CQ or EQ has been overrun.  In such cases, existing
events _may_ be retrievable (or they could be corrupted).
However, when attempting to read from a queue where
events may have been lost or corrupted, report an
overrun error.  The queue is considered to be in an
unrecoverable fatal state at that point.

Fixes #1723

Signed-off-by: Sean Hefty <sean.hefty@intel.com>