prov/psm,psm2: Handle FI_SOURCE in fi_getinfo()
When FI_SOURCE is set, allows HCA unit and port selection be passed via
the "node" argument and be saved to the "src_addr" field of fi_info for
later use in local operations. The address should not be used as the
input of fi_av_insert().

The format of "node" is expected to be "<name>[:<unit>[:<port>]]". A
simple "<name>" implies automatic selection of unit and port.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>