prov/verbs: fix inititialization on several devices
The patch fixes the case then there are several devices in the system
and IPoIB is not configured for all ones. If user tried to initialize
fi_info from other device, it's failing on connection establishment
phase because of verbs context mismatch.

The fix adds cleanup of all rdm-domains which aren't binded to some
interface or explicite defined interface (via env variable)

Signed-off-by:Evgeny Leksikov <evgeny.leksikov@intel.com>