fabric: Add multicast support to the interface
Fixes #1001 - multicast support was deferred until we could
determine the best way to add it without impacting non-multicast
traffic.

Multicast is specified using a new capability bit (FI_MULTICAST).
Multicast groups are described using a new fid type (fid_mc).  A
fid_mc represents a link between an endpoint and a multicast
group.  Each fid_mc contains an fi_addr_t value that is used
to send messages to the multicast group.

In order to determine if the address passed into send() is a
multicast address or an address from an AV, the FI_MULTICAST
flag must be specified with any multicast operation.

Join operations complete asynchronously, with the results
reported to an EQ.  Closing a multicast group results in the
endpoint leaving the group.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>