man/fi_info: Specify limits of src address
The source address returned in fi_info in response to fi_getinfo
is needed when allocating an endpoint.  But it is not guaranteed
to be a usable addresses that can be passed to a remote side. For
example, some of the fields may be a wildcard value.

Similarly, the dest_addr is only usable locally.

Fixes #2475

Signed-off-by: Sean Hefty <sean.hefty@intel.com>