man/endpoint: Relax meaning of total_buffered_recv
Fixes #1240.

Total buffered receive is not well defined, and any attempt to
clarify its meaning will result in forcing a specific implementation.
The buffer size could be defined such that a single large message
would consume the entire buffer, or the size could refer to the
total amount of buffering provided for a set of small messages.
(For example, if the provider implements a rendezvous protocol, then
a single large message would not consume the entire buffer at the
receive side).

Redefine total_buffered_recv to be used as a hint to the provider
as to the amount of unexpected messages that it may need to handle.
It is up to the provider to interpret or use this value as needed.
The FI_RM_ENABLED mode bit should be used by applications in place
of relying on this field.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>