fabric: Add FI_RESTRICTED_COMP mode bit
Fixes #1394.
Fixes #2292.

Some endpoints (such as those supporting tagged operations) may
require FI_CONTEXT, while others may not. However, this can result
in a difficulty for providers when processing completions. A provider
may not know if a completion is associated with an FI_CONTEXT or not.
As a result, the provider could be forced to always require a context
from the app.

The solution is two-fold.  First document that CQs and counters can
only be shared among the same types of endpoints.  This restriction
can be relaxed later, if needed.  Current apps will be unaffected
by this change, as its a documentation modification only, and existing
providers will not need to adhere to this requirement/suggestion.

The second part of the solution is to add a new mode bit that can
be used by applications to indicate that they will only use a CQ
or counter in a specific way: that all attached EPs will have like
attributes.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>