core: remove use of clock_gettime(3)
The portability of clock_gettime(3) on Mac OS / OS X
is... complicated.  Per
https://github.com/Homebrew/homebrew-core/pull/6551 and
https://github.com/ofiwg/libfabric/pull/2508, remove the use of
clock_gettime(3) from the core of libfabric.  clock_gettime(3) is
still used in some providers that do not compile on OS X / MacOS, but
that's ok.

This issue was discovered after v1.4.0 was released and was fixed on
master.  However, we later branched from the v1.4.0 tag to make the
v1.4.x branch in order to fix some specific bugs.  We then released
v1.4.1.

Unfortunately, the fix for this issue was forgotten / not included in
the v1.4.1 release (see
https://github.com/Homebrew/homebrew-core/pull/10292).  Although it is
unlikely, I'm cherry-picking this fix from master to the v1.4.x branch
just in case we ever do a v1.4.2 release.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>

(cherry picked from commit 0b0c8891bbfeb97b6ef8be8e786cf6a990977df8)