fabric: Add comp_list field to fabric attribute
Fixes #2345

Add a new provider component field to the fabric attribute.  This will be
used by utility providers when layering over core providers.  The field
is a character string that is similar to the provider selection
environment variable.

It can be used to describe software layering that is occurring, as well
as which specific implementations may be in use.

The list is a comma separated string that lists which component to
include or exclude, if a component is preceded by '^'.

This allows utility providers to check that they are not creating
circular dependencies, and provides debugging support to show which code
paths are being used.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>