fabric: Add support for job keys
Allow endpoints to be associated with an authorization key, commonly
referred to as a job key.  The auth_key optionally allows an endpoint
to restrict what traffic it accepts.  Only peers with a valid auth_key
can target a restricted endpoint.

Fixes #2529

Architecturally, each endpoint can be associated with a single auth_key.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>