travis: install libibverbs-dev for usnic
This allows the usnic provider to be built.

Unfortunately, the verbose of the librdmacm-dev package in "precise"
Ubuntu is fairly old -- it does not contain <rdma/rsocket.h>.
Meaning: it can't be used to build the verbs provider.  FWIW, newer
versions of Ubuntu have newer versions of librdmacm(-dev) which
contain <rdma/rsocket.h>.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>