travis: for linux clang builds, use -Werror
Somehow, autoconf thinks that the gcc in the Ubuntu "precise" image
that we use in Travis builds does not support C99 (i.e., the
AC_PROG_CC_C99 test comes back with "unsupported").  Hence, when we
use C99 constructs without the --std=c99 flag on the command line, it
results in a warning, and therefore -Werror fails.

Hence, we only use -Werror with clang on Linux.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>