prov/gni: Fix NIC reference issue on MR
In a multi-threaded environment, it is possible that the NIC which
was allocated for the MR would not be the NIC picked by the memory
registration subsystem. To fix this, the NIC reference is taken later
during the registration step.

upstream merge of ofi-cray/libfabric-cray#1103

Signed-off-by: James Swaro <jswaro@cray.com>
(cherry picked from commit ofi-cray/libfabric-cray@7b6c0db24546805baee517de20247072631e4d8c)