dgram_waitset.c: Fix a race condition.
Before this change, the receive for the message was posted by the client
as the final step of the initialization process. The server can
sometimes finish initialization and post the send containing the
message before the client has completed posting a receive buffer. Post
an extra receive buffer to handle the message and synchronize to allow
the client to complete posting the buffer.

Signed-off-by: Ben Turrubiates <bturrubi@cisco.com>