prov/util: Rework util_mr.c
Re-write the util_mr.c code.  Update the code to adhere to the
coding guidelines and preferred variable naming.

Rename ofi_util_mr to ofi_mr_map to better reflect the intent
of the structure as a mapping not a MR.  Simplify the code and
remove unnecessary checks (which should be handled earlier in
the code paths).  Limit the code to handling a single SGE, since
the algorithms assume this.  Add a couple of small optimizations
around structure allocation and initialization.

Fixes #2637

Signed-off-by: Sean Hefty <sean.hefty@intel.com>