prov/rxd: Cleanup of provider code
The progress thread spins, resulting in failures of nearly all
fabtests when run on virtual box.  To fix this, switch from auto
to manual progress.  The rxd cq and ep code is moved to make
better use of the util cq and ep code.  This allows the util
progress code to be used in conjunction with the rxd code.

Apply several code cleanups as part of restructuring the
provider code

Signed-off-by: Sean Hefty <sean.hefty@intel.com>