prov/sockets: read value of FI_SOCKETS_PE_AFFINITY for each new OFI domain
This patch allows to select a different binding for the progression
thread every-time a new domain is opened.
This is particularly important in case several domains with different
NUMA affinities are successively opened.

Also dup the string provided in argument to sock_thread_set_affinity()
as strtok_r() will modify it.

Signed-off-by: Sylvain Didelot <sdidelot@ddn.com>