Fix unbalanced tx/rx counter refs in sockets prov
The sockets provider was adding each counter once per ctx, in spite of
actually referencing it once per event class on the ctx.  As a result,
when closing the ctx, if the same cntr was handling multiple event
classes, it would be removed from the ctx more times than it was added,
causing the ref count on the cntr to become negative (and by extension,
causing an error when the application tried to close the cntr).

This fix adds the cntr on the ctx once per event class for parity with
the operations performed when closing the ctx.

Signed-off-by: James Dinan <james.dinan@intel.com>