usnic: only -libverbs when necessary
Additionally, tighten up the logic for when we -libverbs.
Specifically, only -libverbs when we need it (i.e., when both building
the usnic provider as a DL *and* when we're also building the fake
usnic verbs driver).

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>
(cherry picked from commit 4cf476c81239fcbfedbc31f66d3cbcbb80f9a2d6)