unit/av_test: Remove test with incorrect assumptions
The av_goodbad_2vector_async test assumes that the provider
will process insertions in the same order that they were given
when handling requests asynchronously.  Because the provider
is processing the address asynchronously, no ordering can be
assumed.  Additionally, once a provider fails to insert any
address, it can abort the entire operation, and cancel the
pending insertions.  (This is often desirable from the viewpoint
of the app, since insertions can take significant time, and one
failure may be fatal to the execution of the app).

As a result of these issues, it's safer to simply remove the
test, as it must be entirely re-written to handle these test
cases.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>