Add capability to set default auth key for domain
This commit adds the ability to set the default authorization key
to associate with endpoints and memory registrations created within
a given domain.

If default_auth_keylen is set to 0, then providers will use their
own internal authorization key (if any) as the default authorization
key to associate with endpoints and memory registrations.

Signed-off-by: James Swaro <jswaro@cray.com>