prov/verbs: Remove unpaired lock from EQ destroy
fi_ibv_eq_close() acquires a lock to protect a list that is
being cleaned up.  However, the lock is never released, and both
the lock and list are about to be freed.  So the lock itself is
useless.  Fix the missing lock release by never acquiring it in
the first place.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>