fabric: Rework triggered operations as deferred WQ
In order to support primitives needed to implement collective
operations, re-work the triggered operation definitions.  Triggered
operations are really just deferred work requests.  Instead of
implicitly associating the defered work queue with an endpoint,
associate the WQ with the domain.  This will enable more flexibility
with respect to which type of operations can be queued to the WQ.

Define a mechanism for canceling a specific operations from the WQ,
as well as the ability to flush the WQ can cancel all outstanding
requests.

The new definition will apply when the domain capability is enabled
with FI_TRIGGER support.

Fixes #2630.  Fixes #2558.  Fixes #2559.

This feature will be marked as experimental.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>