fabric/cq-eq: Have application provide error buffer
The current CQ/EQ document that the err_data referenced by
error events points to a provider owned data buffer.  This
results in serialization issues to the application.  Allow
the application to provide a data buffer for error data,
including the size of the input buffer.  Add a domain
attribute to report the maximum size of error data that a
provider may need.

Fixes #2720

Signed-off-by: Sean Hefty <sean.hefty@intel.com>