fabric: Introduce FI_MR_RMA_EVENT mode bit
MRs associated with RMA events may require that the app fully
configure the MR prior to its use.  This handles the case where
the provider cannot dynamically assign a counter to a MR while
it is actively being used.

Such MRs are considered disabled on creation.  Once configured
they must explicitly be enabled (similar to enabling an EP).

Signed-off-by: Sean Hefty <sean.hefty@intel.com>