getinfo: Set FI_PROVIDER before calling no-hint tests
Hints are not passed to the first five tests in getinfo_test.c.
When the test was invoked with a provider argument, this can lead
to the case where errors are reported that are not related to the
provider specified by the provider argument. To avoid this problem,
the environment variable FI_PROVIDER is set before the five no-hint
tests are run. It is then restored to its original value and
the remainder of the tests are run.

Signed-off-by: Jerome Berryhill <Jerome.Berryhill@Intel.com>
Signed-off-by: Sean Hefty <sean.hefty@intel.com>