prov/util: Add api version to ofi_check_info
In order to validate attributes correctly, we need to know
what version of the api the application is attempting to use.

Start by updating ofi_check_info to take the api_version that
was specified by the application.  Update providers that make
use of this call, and fix-up the version being used.  Follow
on patches will push the api_version through to other calls.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>