man/domain: Detail interaction between resouce mgmt and selective comp
There is a potential issue with providers attempting to deal with
operations that should not generate a completion.  I.e. the ep is
configured with FI_SELECTIVE_COMPLETION, with the operation not
generating one.  In the case of an error, the operation actually
should generate a completion.

If FI_RM_ENABLED is set, the provider is required to prevent
overruning the local completion queues.  If an operation will
not generate a completion, it's unclear if the provider must
reserve a CQ entry in the chance an error occurs.  Additionally,
if a CQ entry is reserved, it's unclear how the provider
releases the entry in the success case, since no completion
is actually generated.

Add documentation to the man page to describe this situation
and the expectations on the application.  The solution assumes
that the performance of using selective completions is the
higher priority for how the provider must handle this.

Fixes #2775

Signed-off-by: Sean Hefty <sean.hefty@intel.com>