fabric: Add support for 'shadow' utility provider
Utility providers layer over core providers and have the
following features:

- Uility fabric names match that of the core
- Utility domain names match that of the core
- The provider name is a list of core[;utility]
- Apps that specify a utility provider name to fi_getinfo
  will have their requests directed to that utility provider.
- If a core provider name is also specified, the utility
  will use the specified core provider.
- Apps that specify only a core provider name will have
  their request directed to all utility providers that
  are capable of layering over that core provider.
- The provider version reported by fi_getinfo is the
  topmost provider that is used.
- Utility providers can only layer over core providers.
- Utility providers have a provider name starting with 'ofi-'
- The fi_getinfo prov_attr_only reports utility providers
  directly with no layering.
- Utility providers are not affected by provider filtering.
  I.e. FI_PROVIDER environment variable
- Utility providers always have their logging enabled.

From the viewpoint of an application, the above means that the
fabric attribute prov_name is treated as a substring, rather than
a matching string.  The man pages do not specify how the prov_name
is used when passed as hints.

As part of this change, serveral utility helper functions were
renamed, with parameters renamed as well.  This should make it
easier to understand what the functions are doing and how they
should be used.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>