fabric: Add comment explaining api_version.
When interfacing between ABI 1.0 and ABI 1.1, the api_version field of
the intermediate fi_fabric_attr structure is set to FI_VERSION(1, 0).
This is because there is no versioning information available in the ABI
1.0 call. This of course means that the provider will get a fabric_attr
marked as requesting api version FI_VERSION(1, 0), but the application
could actually be using any version between FI_VERSION(1, 0) and
FI_VERSION(1, 4).

Signed-off-by: Ben Turrubiates <bturrubi@cisco.com>