prov/psm2: Fix an RMA progress bug in multithreaded runs
The provider has an internal call count for fi_cntr_read that is
checked against a threshold so that explicit PSM2 calls can be made
from time to time. In multithreaded runs, it is possible that the
call count is updated more than once to surpass the threshold value
and thus all future checks will fail. As the result no explicit PSM2
progress is made and the performance can become very poor.

Fix the issue by changing the way the threshold is checked against.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>