core: Fix error not calling core provider
When an application calls fi_getinfo with the provider name set
to the util + core (e.g. "ofi-rxm;verbs"), the getinfo call will
be directed to the correct utility provider.  However, when the
utility calls back into fi_getinfo to get information about the
core provider, the call will fail.

The problem is that fi_getinfo splits the provider name into
util and core names.  If a util name is provided, then the getinfo
call will attempt to route back to the util provider.  But the
util provider has set OFI_CORE_PROV_ONLY flag, indicating that
only the core provider should respond.

The solution is to re-order the checks, so that core providers
are called whenever OFI_CORE_PROV_ONLY is set, even if a util
provider name has been specified.  This will allow the util
providers to layer over a core provider properly.

Signed-off-by: Dmitry Gladkov <dmitry.gladkov@intel.com>
Signed-off-by: Sean Hefty <sean.hefty@intel.com>