fabric: Add new FI_CONTEXT2 mode bit
The current fi_context structure is too small for some providers
(e.g. BGQ).  Introduce a new mode bit that doubles the size of
the fi_context structure.

In order to ensure backwards compatibility, the current fi_context
structure is left unchanged.  This avoids the possibility that an
app may have embedded fi_context into a structure optimized to fit
into a specific amount of memory, such as a single cache line.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>