prov/sockets: Add support for cascading counters
The experimental deferred work queue defines a mechanism
to update a counter in response to another counter crossing
a threshold.

Add an implementation of this to the socket provider.  Note
that until we have a test app, we can't verify that this is
correct.  So the code is compile tested only.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>