prov/bgq: Incorrect posting of errors in manual progress mode
The function fi_bgq_cq_enqueue_err was written to propely handle queuing of errors
appropriately as a linked list for manual mode or atomic fifo for auto mode.  However
there were places in the code where error events took place and instead of calling
fi_bgq_cq_enqueue_err posted the error on atomic fifo instead, which then was only
correct for auto mode.  In manual mode this resulted in hangs in some error situations,
one of which was for MPI_Probe when the context probed for had not yet arrived.  The
fix is to utilize the fi_bgq_cq_enqueue_err function in all these situations, and also
there was a but in fi_bgq_cq_enqueue_err regarding locks which was resolved.

Signed-off-by: Paul Coffman <pcoffman@anl.gov>