prov/sockets: Update deferred WQ
Update the experimental triggered operation support (i.e.
deferred work queues) to the latest API proposal.  This
includes adding explicit counters to track completions,
which can be used to map to a more advanced 'scheduling'
algorithm.

The changes themselves are untested, as there are no
current fabtests available for this feature yet.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>