util: Provide helper routines to convert binary addresses to string
This function will help support FI_ADDR_STR in providers that support
sockaddr formats.  For consistency, use it with the AV tostr function
and fi_tostr call, so that address strings all follow the same format.

Note that this changes the string output relative to previous versions
of the library.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>