prov/psm2: Try to turn on PSM2 multi-EP support automatically
The PSM2 multi-EP feature (if present) is controlled by the environment
variable PSM2_MULTI_EP which is off by default. This patch tries to set
this variable automatically if the application asks for the scalable
endpoint feature and the variable has not been set yet.

More specifically, when fi_getinfo() is called the first time, the following
condition is checked: "hints->ep_attr->tx_ctx_cnt > 1". In absence of a
dedictaed flag for scalable endpoint support, this is the next best thing
that captures the scalable endpoint requirement.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>