prov/psm,psm2: Add a synchronization point for local name servers
Name server runs as a separate thread. Sometimes it could start a
little bit slow so that when a local endpoint tries to register its
address the name server has not entered the ready state. This would
lead to future name resolution failure when a remote process tries
to get the address of that endpoint.

A synchronization point is added when the name server thread is
started. Within a limited period of time multiple attempts are made
to connect to the local name server. The synchronization point is
inside the name server start routine and thus does not affect the
execution if the name server is turned off.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>