prov/bgq: Store msg data in context for fi_trecvmsg implementation
The immediate data was not making it into the completion queue entry
for the fi_trecvmsg implementation because the msg data was not being
copied into the fi_bgq_context, in the case of MPI_Mprobe this resulted
in the MPI_Status object not correctly having the MPI_SOURCE field set
as the source rank is sent as the immediate data.  This fix copies that
data into the context and the MPI_Mprobe now works correctly.

Signed-off-by: Paul Coffman <pcoffman@anl.gov>