Alpha release v1.5.0a1
Alpha release to provide packages for early testing of version
1.5.0.  This release is primarily for provider developers to
verify that targeted applications will work with the pending
v1.5 API changes.

Update the NEWS.md file, pulling in from the v1.4.x branch.

Describe the major API changes that were added to the 1.5
version.

Bump the version-info number from 3:1:2 to 10:0:9.  Why you ask?
The v1.4.x branch is likely to be maintained as a long term
maintenance branch, as it's the last version prior to introducing
a significant set of API changes.  Although the 1.4 series _should_
only increment the revision (middle number), we want to keep
open the possibility that more significant changes could be
added to the 1.4 branch without introducing overlapping versions
with the 1.5+ releases.  10:0:9 will show up as .so.1.9.0,
meaning that it supports all apps written to version 1.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>