prov/gni: Release resources prior to test exit
In the RMA tests, it is possible that some resources may not be
released until the err cq is read. This commit changes the tests
to read the error cqes so that the errors are removed.

upstream merge of ofi-cray/libfabric-cray#1370

Signed-off-by: James Swaro <jswaro@cray.com>
(cherry picked from commit ofi-cray/libfabric-cray@abaa5f8aec8f22b169004bc64cbd63ece1915979)